+++ sky-server/src/main/java/com/sky/mapper/SetmealDishMapper.java
@@ -17,7 +17,7 @@ public interface SetmealDishMapper {
     void deleteBatchBySetmealId(List<Long> setmealIds);
 
     @Select("select * from setmeal_dish where setmeal_id = #{setmealId}")
-    List<SetmealDish> getBySetmealId(Integer setmealId);
+    List<SetmealDish> getBySetmealId(Long setmealId);
 
     @Delete("delete from setmeal_dish where setmeal_id = #{setmealId}")
     void deleteBySetmealId(Long setmealId);
