+++ sky-server/src/main/java/com/sky/service/SetmealService.java
@@ -30,7 +30,7 @@ public interface SetmealService {
      */
     void delete(List<Long> ids);
 
-    SetmealVO getByIdWithDish(Integer id);
+    SetmealVO getByIdWithDish(Long id);
 
     void update(SetmealDTO setmealDTO);
